fix(ibkr): widen volume to i64 and surface real snapshot errors
- MarketDataSnapshot.volume and Quote.volume are now i64 so high-volume
  tickers (AAPL/SPY routinely >2.1B/day) don't silently saturate at
  i32::MAX. Bid/ask/last sizes stay i32 — those are per-quote sizes
  that comfortably fit.
- get_market_data_snapshot's None branch now consults
  subscription.error() before defaulting to Timeout, so disconnect /
  decode failures surface with their real cause instead of being
  reported as 5-second timeouts.
- #[allow(dead_code)] sites now carry explanatory comments matching
  the auto_scanner precedent.

+++ src-tauri/src/ibkr/types/quote.rs
@@ -16,7 +16,7 @@ pub struct Quote {
     /// change and change-percent.
     pub prev_close: Option<f64>,
     /// Cumulative session volume.
-    pub volume: Option<i32>,
+    pub volume: Option<i64>,
     /// Unix epoch seconds when the snapshot completed.
     pub timestamp: i64,
 }
